+++ src/backgroundblurprocessor/BackgroundBlurProcessorProvided.ts
@@ -60,6 +60,8 @@ export default class BackgroundBlurProcessorProvided implements BackgroundBlurPr
   protected sourceWidth = 0;
   protected sourceHeight = 0;
   protected blurAmount = 0;
+  protected frameNumber = 0;
+  protected videoFramesPerFilterUpdate = 1;
 
   protected spec: BackgroundFilterSpec;
   protected _blurStrength: number;
@@ -124,6 +126,10 @@ export default class BackgroundBlurProcessorProvided implements BackgroundBlurPr
     if (!options.reportingPeriodMillis) {
       throw new Error('processor has null options - reportingPeriodMillis');
     }
+
+    if (!options.videoFramesPerFilterUpdate) {
+      throw new Error('processor has null options - videoFramesPerFilterUpdate');
+    }
   }
 
   /**
@@ -139,6 +145,7 @@ export default class BackgroundBlurProcessorProvided implements BackgroundBlurPr
 
     this.spec = spec;
     this.logger = options.logger;
+    this.videoFramesPerFilterUpdate = options.videoFramesPerFilterUpdate;
 
     this.setBlurStrength(options.blurStrength);
     this.delegate = new BackgroundBlurVideoFrameProcessorDelegate();
@@ -222,7 +229,7 @@ export default class BackgroundBlurProcessorProvided implements BackgroundBlurPr
   }
 
   /**
-   * This method initializes all of the resource necessary to processs background blur. It returns
+   * This method initializes all of the resource necessary to processes background blur. It returns
    * a promise and resolves or rejects the promise once the initialization is complete.
    * @returns
    * @throws An error will be thrown
@@ -303,8 +310,8 @@ export default class BackgroundBlurProcessorProvided implements BackgroundBlurPr
 
       if (this.scaledCanvas === undefined) {
         this.scaledCanvas = document.createElement('canvas');
-        this.scaledCanvas.width = inputCanvas.width * hscale;
-        this.scaledCanvas.height = inputCanvas.height * vscale;
+        this.scaledCanvas.width = this.spec.model.input.width;
+        this.scaledCanvas.height = this.spec.model.input.height;
       }
 
       const scaledCtx = this.scaledCanvas.getContext('2d');
@@ -320,17 +327,20 @@ export default class BackgroundBlurProcessorProvided implements BackgroundBlurPr
         this.scaledCanvas.height
       );
 
-      // process frame...
-      const maskPromise = this.mask$.whenNext();
-      this.worker.postMessage({ msg: 'predict', payload: imageData }, [imageData.data.buffer]);
-
-      mask = await maskPromise;
+      // update the filter mask based on the filter update rate
+      if (this.frameNumber % this.videoFramesPerFilterUpdate === 0) {
+        // process frame...
+        const maskPromise = this.mask$.whenNext();
+        this.worker.postMessage({ msg: 'predict', payload: imageData }, [imageData.data.buffer]);
+        mask = await maskPromise;
+      }
       this.drawImageWithMask(inputCanvas, mask);
     } catch (error) {
       this.logger.error(`could not process background blur frame buffer due to ${error}`);
       return buffers;
     } finally {
       this.frameCounter.filterComplete();
+      this.frameNumber++;
     }
 
     buffers[0] = this.canvasVideoFrameBuffer;